Login/Signup Added
The SQLite3 database and superuser are set up. Restaurant managers can sign up, and the signup page leads to the login page. The login page checks the user's role—if it's a restaurant manager, they go to the restaurant homepage; if it's a warehouse manager, they go to the warehouse homepage. Still need to add a check to make sure a restaurant manager can only continue if their status is approved.